Refactorando controller
